+++ fsw/pc-rtems/inc/psp_version.h
@@ -27,7 +27,7 @@
 /*
  * Development Build Macro Definitions
  */
-#define CFE_PSP_IMPL_BUILD_NUMBER   67
+#define CFE_PSP_IMPL_BUILD_NUMBER   72
 #define CFE_PSP_IMPL_BUILD_BASELINE "v1.6.0-rc4"
 
 /*